app/src/androidTest/java/com/udacity/bakingapp/MainActivityTest.java
package com.udacity.bakingapp;

import android.support.test.espresso.IdlingRegistry;
import android.support.test.espresso.IdlingResource;
import android.support.test.espresso.matcher.ViewMatchers.Visibility;
import android.support.test.rule.ActivityTestRule;
import android.support.test.runner.AndroidJUnit4;
import android.support.v7.widget.RecyclerView;

import org.junit.After;
import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;

import static android.support.test.espresso.Espresso.onView;
import static android.support.test.espresso.action.ViewActions.click;
import static android.support.test.espresso.assertion.ViewAssertions.matches;
import static android.support.test.espresso.contrib.RecyclerViewActions.actionOnItemAtPosition;
import static android.support.test.espresso.matcher.ViewMatchers.assertThat;
import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
import static android.support.test.espresso.matcher.ViewMatchers.withEffectiveVisibility;
import static android.support.test.espresso.matcher.ViewMatchers.withId;
import static android.support.test.espresso.matcher.ViewMatchers.withText;
import static org.hamcrest.Matchers.is;

@RunWith(AndroidJUnit4.class)
public class MainActivityTest {

    @Rule
    public ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);

    private IdlingResource mIdlingResource;

    @Before
    public void registerIdlingResource() {
        mIdlingResource = mActivityTestRule.getActivity().getIdlingResource();
        IdlingRegistry.getInstance().register(mIdlingResource);
    }

    @After
    public void unregisterIdlingResource() {
        if (mIdlingResource != null) {
            IdlingRegistry.getInstance().unregister(mIdlingResource);
        }
    }

    @Test
    public void testLoadRecipes() {
        onView(withId(R.id.tv_error_message_display))
                .check(matches(withEffectiveVisibility(Visibility.INVISIBLE)));
        // Check adapter item count
        onView(withId(R.id.recyclerview))
                .check(matches(withEffectiveVisibility(Visibility.VISIBLE)))
                .check((view, noViewFoundException) ->
                        assertThat(((RecyclerView) view).getAdapter().getItemCount(), is(4)));
    }

    @Test
    public void testClickRecipe() {
        onView(withId(R.id.recyclerview))
                .perform(actionOnItemAtPosition(1, click()));

        // Check that title is displayed
        onView(withText("Brownies"))
                .check(matches(isDisplayed()));
    }
}
